// frontend/src/client/types.gen.ts
export type MessagePublic = {
    id: string;
    sender: string;
    sender_name: string;
    sender_last_name: string;
    receiver: string;
    receiver_name: string;
    receiver_last_name: string;
    message: string;
    sender_profile_image: (string | null);
    receiver_profile_image: (string | null);
    sent_at: string;
};